requirements.txt
requests
flask
flask-cors
gunicorn
python-dotenv
psycopg2-binary
bcrypt
PyJWT
